hemoglobin as its own component
