+++ src/vivarium_gates_iv_iron/components/__init__.py
@@ -11,3 +11,4 @@ from vivarium_gates_iv_iron.components.pregnancy import Pregnancy
 from vivarium_gates_iv_iron.components.children import BirthRecorder
 from vivarium_gates_iv_iron.components.maternal_bmi import MaternalBMIExposure
 from vivarium_gates_iv_iron.components.intervention import MaternalInterventions
+from vivarium_gates_iv_iron.components.hemoglobin import Hemoglobin
